Fix/Wrap all access to the global_context_info_ shared pointer in atomic store/load functions
The actual ContextInfo that we point at is immutable (const), so there's no risk
in accessing it from multiple threads.
However, it is *not* thread-safe to have concurrent access to the *pointer address*
held in the global_context_info_ shared pointer, since the address itself might be
modified by a different thread if replace/update global_context_info is called.

Theoretically (to my understanding), if we copy/access the pointer held
by the shared_ptr while it is being concurrently replaced by another
thread, we might end up working on a corrupted address. Using atomic
store/load operations protects from this risk.

See:
https://stackoverflow.com/questions/40223599/what-is-the-difference-between-stdshared-ptr-and-stdatomicstdshared-ptr
https://en.cppreference.com/w/cpp/memory/shared_ptr/atomic
https://ddanilov.me/usage-of-shared_ptr