Comments (note that context_info keys could be duplicated when logging to console)
Note that in the case of formatted_log_plaintext_long (i.e when logging to the console
with LineFormat::PLAINTEXT_LONG) then any duplicate keys in the different context_infos
will be logged multiple times.
For example: If a given key (e.g "bla_bla_id") will be present both in
the `log_context` and in the `logger_context` then it will be logged
twice.
(unlike in the case of other sinks where it will only appear once, and
the most local context will determine the value)