ADB-2328: fix/workaround: Protect the GlobalContextInfo with a ForkSafeMutex instead of using atomic_load/store, which caused deadlocks after forking (#41)
**TL;DR**
This PR should solve the potential deadlocks after forking, when attempting to log.
Instead of using atomic load/store operations on the GlobalContextInfo, we will now protect it with a (fork-safe) mutex. (because the atomic operations on a shared_ptr use a mutex internally, which isn't fork-safe...)
This shouldn't cause any performance degradation, since we are simply replacing an implicit locking (by std::atomic operations) with an explicit locking (by us).
In the future, after upgrading to c++20 we should remove this lock and use `std::atomic_shared_ptr` (which is lock-free).

---

**Details**
We observed occasional deadlocks after forking when attempting to `std::atomic_load` the global ContextInfo:
```
5  0x000055aaed45668e in std::atomic_load_explicit<octo::logger::ContextInfo const> (__p=0x14b18f82a158) at /usr/include/c++/10/bits/shared_ptr_atomic.h:105
6  std::atomic_load<octo::logger::ContextInfo const> (__p=0x14b18f82a158) at /usr/include/c++/10/bits/shared_ptr_atomic.h:112
7  octo::logger::Manager::dump (this=0x14b18f82a0c0, log=..., channel=..., context_info=...) at /home/jenkins/agent/workspace/to-logger-cpp-master-full_master/source/src/manager.cpp:143
```
(see full backtrace and details in [ADB-2328](https://ca-il-jira.il.cyber-ark.com:8443/browse/ADB-2328) and [ADB-2324](https://ca-il-jira.il.cyber-ark.com:8443/browse/ADB-2324))

To our understanding, this is because `std::atomic_load<std::shared_ptr>` uses a mutex internally, and this mutex in not fork-safe.

Instead of using the atomic operations, which *implicitly* lock, we will now *explicitly* lock with our own mutex, which will enable us to be fork-safe (with no real cost to performance, since anyway there was already a lock, it just wasn't called directly by us)

IMPORTANT: The ideal solution is to use `std::atomic_shared_ptr`, which would provide lock-free atomic operations, and enable us to get rid of the `global_context_info_mutex_`, and also be fork-safe, but this requires upgrading to C++20...

Another possible solution could be to switch to fork/execing, which would solve all the fork-safety issues we encounter.

---
**Additional minor changes**

**ForkSafeMutex**
* ctor made `explicit`
* Using `default` dtor (don't see any reason to call `fork_reset()` in the dtor). If the mutex is destructed then there's no more risk of deadlocking and the reset is unnecesssary. (unless I'm missing something?)
* dtor no longer `virtual` (didn't see any reason for it to be)
* Removed the `get()` and `operator*`, and instead added an implicit conversion to the underlying MutexType, so that ForkSafeMutex behaves like a regular mutex.

**Manager**
* Simplify logic in `Manager::instance()`
* ctor made `explicit`
* explicitly delete move operator/ctor